modularized the driver components, seperated the driver implementations into seperate files for ease
