+++ pgsqltoolsservice/driver/__init__.py
@@ -1,0 +1,9 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from pgsqltoolsservice.driver.types import ServerConnection
+from pgsqltoolsservice.driver.driver_manager import DriverManager
+
+__all__ = ['DriverManager', 'ServerConnection']
