+++ pgsqltoolsservice/driver/types/__init__.py
@@ -1,0 +1,10 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from pgsqltoolsservice.driver.types.driver import ServerConnection
+from pgsqltoolsservice.driver.types.psycopg_driver import PsycopgConnection
+from pgsqltoolsservice.driver.types.pymysql_driver import PyMySQLConnection
+
+__all__ = ['ServerConnection', 'PyMySQLConnection', 'PsycopgConnection']
