+++ pgsqltoolsservice/edit_data/smo_edit_table_metadata_factory.py
@@ -9,7 +9,7 @@ from typing import List  # noqa
 from pgsmo import Server
 from pgsmo.objects.table.table import Table  # noqa
 from pgsmo.objects.table_objects.column import Column  # noqa
-from pgsqltoolsservice.utils.driver import ServerConnection
+from pgsqltoolsservice.driver import ServerConnection
 from pgsqltoolsservice.edit_data import EditTableMetadata, EditColumnMetadata
 from pgsqltoolsservice.utils import object_finder
 from pgsqltoolsservice.metadata.contracts.object_metadata import ObjectMetadata
